MapDataStore: remove MapReadResultBuilder, #668
