+++ mapsforge-map-android/src/main/java/org/mapsforge/map/android/view/MapView.java
@@ -96,6 +96,7 @@ public class MapView extends ViewGroup implements org.mapsforge.map.view.MapView
 				GRAPHIC_FACTORY, this.model.displayModel);
 	}
 
+	@Override
 	public void addLayer(Layer layer) {
 		this.layerManager.getLayers().add(layer);
 	}
@@ -245,6 +246,7 @@ public class MapView extends ViewGroup implements org.mapsforge.map.view.MapView
 		this.mapZoomControls.setShowMapZoomControls(showZoomControls);
 	}
 
+	@Override
 	public void setCenter(LatLong center) {
 		this.model.mapViewPosition.setCenter(center);
 	}
@@ -261,6 +263,7 @@ public class MapView extends ViewGroup implements org.mapsforge.map.view.MapView
 		this.mapScaleBar = mapScaleBar;
 	}
 
+	@Override
 	public void setZoomLevel(byte zoomLevel) {
 		this.model.mapViewPosition.setZoomLevel(zoomLevel);
 	}
